+++ src/main/scala/com/testerhome/appcrawler/Crawler.scala
@@ -350,7 +350,7 @@ class Crawler extends CommonLog {
     URIElement(currentUrl, tag, id, name, loc)
   }
 
-  def isBackApp(): Boolean = {
+  def needBackApp(): Boolean = {
     //跳到了其他app. 排除点一次就没有的弹框
     if (conf.appWhiteList.forall(appNameRecord.last().toString.matches(_)==false) && appNameRecord.last(2).distinct.size>1) {
       log.warn(s"not in app white list ${conf.appWhiteList}")
@@ -363,7 +363,7 @@ class Crawler extends CommonLog {
 
   }
 
-  def isExit(): Boolean = {
+  def needExit(): Boolean = {
     //超时退出
     if ((new Date().getTime - startTime) > conf.maxTime * 1000) {
       log.warn("maxTime out Quit need exit")
@@ -381,7 +381,7 @@ class Crawler extends CommonLog {
     return false
   }
 
-  def isReturn(): Boolean = {
+  def needReturn(): Boolean = {
     //url黑名单
     if (conf.urlBlackList.exists(urlStack.head.matches(_))) {
       log.warn(s"${urlStack.head} in urlBlackList should return")
@@ -437,19 +437,34 @@ class Crawler extends CommonLog {
   }
 
   def isSmall(m: immutable.Map[String, Any]): Boolean ={
+
+    var res=false
+    //todo: support ios
     if(m.getOrElse("bounds", "").toString.nonEmpty){
       val bounds="\\d+".r().findAllIn(m.get("bounds").get.toString).matchData.map(_.group(0)).toList
-      val width=bounds(2).toInt-bounds(0).toInt
-      val height=bounds(3).toInt-bounds(1).toInt
-      //高度小就跳过
-      if(height<50){
-        true
-      }else{
-        false
+      val startX=bounds(0).toInt
+      val startY=bounds(1).toInt
+      val endX=bounds(2).toInt
+      val endY=bounds(3).toInt
+      val width=endX-startX
+      val height=endY-startY
+
+      if(endY>driver.screenHeight || endX>driver.screenWidth){
+        log.info(bounds)
+        log.info(driver.screenHeight)
+        log.info("not visual")
+        res=true
       }
-    }else{
-      false
+
+      //高度小就跳过
+      if(height<50 && width<50){
+        log.info(bounds)
+        log.info(driver.screenHeight)
+        log.info("small")
+        res=true
+      }
     }
+    res
   }
 
   //todo: 支持xpath表达式
@@ -459,10 +474,6 @@ class Crawler extends CommonLog {
     var lastElements = List[immutable.Map[String, Any]]()
     var selectedElements = List[immutable.Map[String, Any]]()
     var blackElements = List[immutable.Map[String, Any]]()
-
-    val allElements = driver.getListFromXPath("//*")
-    log.trace(s"all node size = ${allElements.size}")
-    allElements.foreach(log.trace)
 
     conf.selectedList.foreach(xpath => {
       log.trace(s"selectedList xpath =  ${xpath}")
@@ -475,10 +486,9 @@ class Crawler extends CommonLog {
 
     //remove blackList
     conf.blackList.foreach(xpath => {
-      log.info(s"blackList check xpath = ${xpath}")
       val temp = driver.getListFromXPath(xpath)
       temp.foreach(x=>
-        log.info(s"blackList hit  ${x}")
+        log.info(s"blackList hit ${xpath} ${x}")
       )
       blackElements ++= temp
     })
@@ -488,8 +498,8 @@ class Crawler extends CommonLog {
     selectedElements.map(_.getOrElse("xpath", "no xpath")).foreach(log.trace)
 
     //exclude small
-    selectedElements=selectedElements.filter(isValid)
-    log.info(s"all - small - non valid elements size = ${selectedElements.size}")
+    selectedElements=selectedElements.filter(isSmall(_)==false)
+    log.info(s"all - small elements size = ${selectedElements.size}")
     selectedElements.map(_.getOrElse("xpath", "no xpath")).foreach(log.trace)
 
     //sort
@@ -714,7 +724,7 @@ class Crawler extends CommonLog {
     //判断下一步动作
 
     //是否应该退出
-    if (isExit()) {
+    if (needExit()) {
       log.warn("get signal to exit")
       return
     }
@@ -729,7 +739,7 @@ class Crawler extends CommonLog {
     }
 
     //是否需要回退到app
-    if (isBackApp()) {
+    if (needBackApp()) {
       nextElement = Some(URIElement(s"${currentUrl}", "", "", "",
         s"backApp-${appNameRecord.last()}-${store.clickedElementsList.size}"))
       setElementAction("backApp")
@@ -749,7 +759,7 @@ class Crawler extends CommonLog {
 
     //判断是否需要返回上层
     if (nextElement == None) {
-      if (isReturn()) {
+      if (needReturn()) {
         log.info("need to back")
         nextElement = getBackButton()
         setElementAction("back")
@@ -967,6 +977,11 @@ class Crawler extends CommonLog {
           conf.defaultBackAction.foreach(Runtimes.eval)
         } else {
           driver.backApp()
+          driver.getPageSource()
+          if(needReturn()){
+            driver.launchApp()
+          }
+
         }
       }
       case "monkey" => {