+++ src/test/java/XueqiuDemo.java
@@ -1,0 +1,45 @@
+import io.appium.java_client.MobileElement;
+import io.appium.java_client.android.AndroidDriver;
+import junit.framework.TestCase;
+import org.junit.After;
+import org.junit.Before;
+import org.junit.Test;
+import static org.junit.Assert.*;
+import java.net.MalformedURLException;
+import java.net.URL;
+import java.util.concurrent.TimeUnit;
+
+import org.openqa.selenium.By;
+import org.openqa.selenium.remote.DesiredCapabilities;
+
+public class XueqiuDemo {
+
+    private AndroidDriver driver;
+
+    @Before
+    public void setUp() throws MalformedURLException {
+        DesiredCapabilities desiredCapabilities = new DesiredCapabilities();
+        desiredCapabilities.setCapability("appPackage", "com.xueqiu.android");
+        desiredCapabilities.setCapability("appActivity", ".view.WelcomeActivityAlias");
+        desiredCapabilities.setCapability("platformName", "android");
+        desiredCapabilities.setCapability("deviceName", "ddd");
+
+        URL remoteUrl = new URL("http://localhost:4723/wd/hub");
+
+        driver = new AndroidDriver(remoteUrl, desiredCapabilities);
+    }
+
+    @Test
+    public void sampleTest() throws InterruptedException {
+        //Thread.sleep(12000);
+        driver.manage().timeouts().implicitlyWait(60, TimeUnit.SECONDS);
+        MobileElement el1 = (MobileElement) driver.findElementById("com.xueqiu.android:id/login_account");
+        el1.sendKeys("15600534760");
+        assertTrue(el1.isDisplayed());
+    }
+
+    @After
+    public void tearDown() {
+        driver.quit();
+    }
+}