+++ src/test/scala/com/testerhome/appcrawler/it/TestWebView.scala
@@ -3,10 +3,12 @@ package com.testerhome.appcrawler.it
 import java.net.URL
 
 import com.testerhome.appcrawler.RichData
+import io.appium.java_client.TouchAction
 import io.appium.java_client.android.AndroidDriver
 import org.openqa.selenium.{By, WebElement}
-import org.openqa.selenium.remote.{RemoteWebDriver, DesiredCapabilities}
+import org.openqa.selenium.remote.{DesiredCapabilities, RemoteWebDriver}
 import org.scalatest.FunSuite
+
 import collection.JavaConversions._
 
 /**
@@ -34,15 +36,6 @@ class TestWebView extends FunSuite{
     driver.getContextHandles.toArray.foreach(x=>println(x))
     //driver.context("WEBVIEW_com.ihealthlabs.ijiankang.patient.android")
     val size=driver.manage().window().getSize
-    driver.swipe( (size.width*0.9).toInt, (size.height*0.5).toInt,
-      (size.width*0.1).toInt, (size.height*0.5).toInt, 2000)
-
-    driver.swipe( (size.width*0.9).toInt, (size.height*0.5).toInt,
-      (size.width*0.1).toInt, (size.height*0.5).toInt, 2000)
-
-    driver.swipe( (size.width*0.9).toInt, (size.height*0.5).toInt,
-      (size.width*0.1).toInt, (size.height*0.5).toInt, 2000)
-
     xml=driver.getPageSource
     println(RichData.toPrettyXML(xml))
     driver.getContextHandles.toArray.foreach(x=>println(x))