+++ CHANGELOG.md
@@ -29,7 +29,10 @@ testcase:
     then:
     - //*[contains(@text, "港股")]
 ```
-
+所有的xpath的设置都支持如下三种形式
+- xpath //*[contains(@resource-id, 'ddd')]
+- regex ^确定$ 
+- contains关系 取消 确定
 # 2.0.0
 支持macaca[完成]  
 失败重试[完成]